+++ docs/source/export.rst
@@ -3,11 +3,11 @@ torch.export
 
 .. TODO: Add torch.export() tutorial here.
 
-.. automodule:: torch
-.. autofunction:: export
-
 .. warning::
     This feature is a prototype and may have compatibility breaking changes in the future.
+
+.. automodule:: torch.export
+.. autofunction:: export
 
 .. toctree::
    :glob:
